requirements.txt
numpy~=1.24.1
torchmetrics~=1.5.2
pycocotools~=2.0.7